requirements.txt
streamlit
langchain
langchain_community
langchain-groq
wikipedia
arxiv
tavily-python
shopifyapi
python-dotenv